web/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
python-jose[cryptography]>=3.3
cryptography>=44.0
supabase>=2.0
python-multipart>=0.0.9
Pillow>=10.0
